Move at::{Refcounted,}MapAllocator to c10
libshm.so depends on the torch library exclusively for
`at::RefcountedMapAllocator`, so it makes sense to move it to c10 along with the
other other memory allocators.

This means `libshm.so` only depends on `c10` and we don't need to relink
`libshm.so` for every ATen change.
