+++ BUILD.bazel
@@ -1648,7 +1648,8 @@ cc_library(
     name = "shm",
     srcs = glob(["torch/lib/libshm/*.cpp"]),
     deps = [
-        ":torch",
+        ":torch_headers",
+        "//c10",
     ],
 )
