+++ aten/src/ATen/CMakeLists.txt
@@ -378,28 +378,6 @@ if(UNIX AND NOT APPLE)
    endif(NEED_LIBRT)
 endif(UNIX AND NOT APPLE)
 
-if(UNIX)
-  set(CMAKE_EXTRA_INCLUDE_FILES "sys/mman.h")
-  CHECK_FUNCTION_EXISTS(mmap HAVE_MMAP)
-  if(HAVE_MMAP)
-    add_definitions(-DHAVE_MMAP=1)
-  endif(HAVE_MMAP)
-  # done for lseek: https://www.gnu.org/software/libc/manual/html_node/File-Position-Primitive.html
-  add_definitions(-D_FILE_OFFSET_BITS=64)
-  CHECK_FUNCTION_EXISTS(shm_open HAVE_SHM_OPEN)
-  if(HAVE_SHM_OPEN)
-    add_definitions(-DHAVE_SHM_OPEN=1)
-  endif(HAVE_SHM_OPEN)
-  CHECK_FUNCTION_EXISTS(shm_unlink HAVE_SHM_UNLINK)
-  if(HAVE_SHM_UNLINK)
-    add_definitions(-DHAVE_SHM_UNLINK=1)
-  endif(HAVE_SHM_UNLINK)
-  CHECK_FUNCTION_EXISTS(malloc_usable_size HAVE_MALLOC_USABLE_SIZE)
-  if(HAVE_MALLOC_USABLE_SIZE)
-    add_definitions(-DHAVE_MALLOC_USABLE_SIZE=1)
-  endif(HAVE_MALLOC_USABLE_SIZE)
-endif(UNIX)
-
 ADD_DEFINITIONS(-DUSE_EXTERNAL_MZCRC)
 
 if(NOT MSVC)
