+++ c10/build.bzl
@@ -4,6 +4,7 @@ def define_targets(rules):
         visibility = ["//visibility:public"],
         deps = [
             "//c10/core:CPUAllocator",
+            "//c10/core:MapAllocator",
             "//c10/core:ScalarType",
             "//c10/core:alignment",
             "//c10/core:alloc_cpu",
