+++ c10/core/build.bzl
@@ -19,6 +19,27 @@ def define_targets(rules):
     )
 
     rules.cc_library(
+        name = "MapAllocator",
+        srcs = ["MapAllocator.cpp"],
+        hdrs = ["MapAllocator.h"],
+        linkstatic = True,
+        local_defines = [
+            "C10_BUILD_MAIN_LIB",
+            "HAVE_MMAP=1",
+            "HAVE_SHM_OPEN=1",
+            "HAVE_SHM_UNLINK=1",
+        ],
+        visibility = ["//visibility:public"],
+        deps = [
+            ":CPUAllocator",
+            ":base",
+            "//c10/util:base",
+        ],
+        linkopts = ["-lrt"],
+        alwayslink = True,
+    )
+
+    rules.cc_library(
         name = "ScalarType",
         hdrs = ["ScalarType.h"],
         linkstatic = True,
@@ -61,6 +82,7 @@ def define_targets(rules):
             ],
             exclude = [
                 "CPUAllocator.cpp",
+                "MapAllocator.cpp",
                 "impl/alloc_cpu.cpp",
             ],
         ),
@@ -71,6 +93,7 @@ def define_targets(rules):
             ],
             exclude = [
                 "CPUAllocator.h",
+                "MapAllocator.h",
                 "impl/alloc_cpu.h",
             ],
         ),
