+++ torch/lib/libshm_windows/core.cpp
@@ -10,17 +10,17 @@ static void deleteTHManagedMapAllocator(void* ptr) {
   delete static_cast<THManagedMapAllocator*>(ptr);
 }
 
-at::DataPtr THManagedMapAllocator::makeDataPtr(
+c10::DataPtr THManagedMapAllocator::makeDataPtr(
     const char* manager_handle,
     const char* filename,
     int flags,
     size_t size) {
   auto* context =
       new THManagedMapAllocator(manager_handle, filename, flags, size);
-  return {context->data(), context, &deleteTHManagedMapAllocator, at::kCPU};
+  return {context->data(), context, &deleteTHManagedMapAllocator, c10::kCPU};
 }
 
 THManagedMapAllocator* THManagedMapAllocator::fromDataPtr(
-    const at::DataPtr& dptr) {
+    const c10::DataPtr& dptr) {
   return dptr.cast_context<THManagedMapAllocator>(&deleteTHManagedMapAllocator);
 }
