+++ torch/lib/libshm_windows/libshm.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ATen/MapAllocator.h>
+#include <c10/core/MapAllocator.h>
 
 #ifdef __cplusplus
 
@@ -12,21 +12,21 @@
 
 SHM_API void libshm_init(const char* manager_exec_path);
 
-class SHM_API THManagedMapAllocator : public at::RefcountedMapAllocator {
+class SHM_API THManagedMapAllocator : public c10::RefcountedMapAllocator {
  public:
   THManagedMapAllocator(
       const char* manager_handle,
       const char* filename,
       int flags,
       size_t size)
-      : at::RefcountedMapAllocator(filename, flags, size) {}
+      : c10::RefcountedMapAllocator(filename, flags, size) {}
 
-  static at::DataPtr makeDataPtr(
+  static c10::DataPtr makeDataPtr(
       const char* manager_handle,
       const char* filename,
       int flags,
       size_t size);
-  static THManagedMapAllocator* fromDataPtr(const at::DataPtr&);
+  static THManagedMapAllocator* fromDataPtr(const c10::DataPtr&);
 
   const char* manager_handle() const {
     return "no_manager";
